fix: use dtype=float in CountVectorizer

+++ app.py
@@ -50,8 +50,8 @@ def _train():
                 if isinstance(ex, str) and ex.strip():
                     X_texts.append(normalize(ex))
                     y_labels.append(intent)
-        vec = CountVectorizer(analyzer="char_wb", ngram_range=(3, 5))
-        X_v = vec.fit_transform(X_texts).astype("float32")
+        vec = CountVectorizer(analyzer="char_wb", ngram_range=(3, 5), dtype=float)
+        X_v = vec.fit_transform(X_texts)
         clf = MLPClassifier(hidden_layer_sizes=(64,), max_iter=200, random_state=42, early_stopping=True)
         clf.fit(X_v, y_labels)
         _vec, _X_v, _clf, _cfg = vec, X_v, clf, cfg
